requirements.txt
# Flask core
Flask>=3.0.0
Werkzeug>=3.0.0
Flask-Login>=0.6.3
gunicorn>=22.0.0

# Database
PyMySQL>=1.1.0
psycopg[binary]>=3.1.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
alembic>=1.13.0

# Environment & config
python-dotenv>=1.0.0

# API docs + validation
pydantic>=2.0.0
spectree>=1.4.0

# Password hashing
argon2-cffi>=23.1.0

# Scheduling (optional nightly retrain)
APScheduler>=3.10.0

# ML / forecasting
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
prophet>=1.1.0
mlflow>=2.10.0
holidays>=0.40

# LLM (OpenRouter — OpenAI-compatible client)
openai>=1.0.0
